Classic Versions Removal
Going to try extending this project a bit, going to start fresh to some extent :)

I love the web too, I think this could be a very awesome app! Let alone it already working very well, I think it could be modernized a bit in terms of the codebase and the bundling setup.